[FIX] Modify Save and Discard Button, Type change in Company Metrics

+++ custom_addons/incident_management/models/x_company_monthly_metrics.py
@@ -35,7 +35,7 @@ class MonthlyMetrics(models.Model):
 
     working_days = fields.Integer(string='Working Days')
     hour_worked = fields.Float(string='Hours Worked')
-    total_sales = fields.Integer(string='Total Sales')
+    total_sales = fields.Float(string='Total Sales')
 
     @api.depends('month_selection', 'year_selection')
     def _compute_name(self):
